fix(src/pulsar/service): remove metric from here and add the metric to the right place
